+++ packages/@tailwindcss-upgrade/src/migrate-js-config.ts
@@ -1,6 +1,7 @@
+import { Scanner } from '@tailwindcss/oxide'
 import fs from 'node:fs/promises'
 import { dirname } from 'path'
-import type { Config } from 'tailwindcss'
+import { type Config } from 'tailwindcss'
 import defaultTheme from 'tailwindcss/defaultTheme'
 import { fileURLToPath } from 'url'
 import { loadModule } from '../../@tailwindcss-node/src/compile'
@@ -15,7 +16,7 @@ import type { ThemeConfig } from '../../tailwindcss/src/compat/config/types'
 import { darkModePlugin } from '../../tailwindcss/src/compat/dark-mode'
 import type { DesignSystem } from '../../tailwindcss/src/design-system'
 import { findStaticPlugins, type StaticPluginOptions } from './utils/extract-static-plugins'
-import { info } from './utils/renderer'
+import { info, warn } from './utils/renderer'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
@@ -54,7 +55,7 @@ export async function migrateJsConfig(
   }
 
   if ('content' in unresolvedConfig) {
-    sources = migrateContent(unresolvedConfig as any, base)
+    sources = await migrateContent(unresolvedConfig as any, base)
   }
 
   if ('theme' in unresolvedConfig) {
@@ -158,16 +159,35 @@ function createSectionKey(key: string[]): string {
   return sectionSegments.join('-')
 }
 
-function migrateContent(
+async function migrateContent(
   unresolvedConfig: Config & { content: any },
   base: string,
-): { base: string; pattern: string }[] {
+): Promise<{ base: string; pattern: string }[]> {
+  let autoContentFiles = listAutoContentFiles(base)
+
   let sources = []
   for (let content of unresolvedConfig.content) {
     if (typeof content !== 'string') {
       throw new Error('Unsupported content value: ' + content)
     }
-    sources.push({ base, pattern: content })
+
+    let sourceFiles = listSourceContentFiles({ base, pattern: content })
+
+    let autoContentContainsAllSourceFiles = true
+    for (let sourceFile of sourceFiles) {
+      if (!autoContentFiles.includes(sourceFile)) {
+        autoContentContainsAllSourceFiles = false
+        break
+      }
+    }
+
+    if (autoContentContainsAllSourceFiles) {
+      warn(
+        'The `content` configuration `${content}` is already included in the automatic content file discovery and will not be migrated.',
+      )
+    } else {
+      sources.push({ base, pattern: content })
+    }
   }
   return sources
 }
@@ -252,4 +272,16 @@ function onlyAllowedThemeValues(theme: ThemeConfig): boolean {
 function keyframesToCss(keyframes: Record<string, unknown>): string {
   let ast: AstNode[] = keyframesToRules({ theme: { keyframes } })
   return toCss(ast).trim() + '\n'
+}
+
+function listAutoContentFiles(base: string) {
+  let scanner = new Scanner({ detectSources: { base } })
+  scanner.scan()
+  return scanner.files
+}
+
+function listSourceContentFiles(source: { base: string; pattern: string }): string[] {
+  let scanner = new Scanner({ sources: [source] })
+  scanner.scan()
+  return scanner.files
 }