+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - _Upgrade (experimental)_: Migrate `plugins` with options to CSS ([#14700](https://github.com/tailwindlabs/tailwindcss/pull/14700))
 
+### Changed
+
+- _Upgrade (experimental)_: `content` rules from the JS configuration that are also covered by the automatic source detection are no longer migrated to CSS ([#14714](https://github.com/tailwindlabs/tailwindcss/pull/14714))
+
 ## [4.0.0-alpha.28] - 2024-10-17
 
 ### Added